functions/requirements.txt
Flask
requests
google-cloud-translate